Remove ParisHackers
Hello! I'm the original creator of ParisHackers, which indeed started as a HN meetup that ran for a couple years. However after a job change I lost access to the meetup group, which was subsequently not renewed and taken over by some crypto people with absolutely no interest in HN. All this is quite obvious from the meetup history so it should not appear in this list anymore. thanks!